+++ CHANGELOG.md
@@ -54,3 +54,4 @@
 - [x] plugin prometheus [prometheus](https://apisix.apache.org/zh/docs/apisix/plugins/prometheus/) 2024-04-21
 - [x] standalone mode file watcher 2024-04-22
 - [x] global rules 2024-04-23
+- [x] plugin_config_id in route 2024-04-28